requirements.txt
a2wsgi==1.7.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.17.0
app-commands==0.1.0
astroid==2.8.0
Async-OAuthlib==0.0.9
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
beautifulsoup4==4.12.2
better-ipc==2.0.3
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
DateTime==4.7
discord-ext-ipc==2.1.1
discord.py==2.3.2
distro==1.6.0
dominate==2.6.0
EasyProcess==1.1
entrypoint2==1.1
ffmpeg==1.4
ffmpeg-python==0.2.0
Flask==3.0.0
frozenlist==1.4.0
future==0.18.3
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
isort==5.9.3
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.1.2
lazy-object-proxy==1.6.0
MarkupSafe==2.1.3
mccabe==0.6.1
MouseInfo==0.1.3
mss==9.0.1
multidict==6.0.4
oauthlib==3.2.2
packaging==21.0
parso==0.8.2
Pillow==10.0.1
pkgconfig==1.5.5
platformdirs==2.4.0
prettytable==3.9.0
priority==2.0.0
psutil==5.9.2
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg-pool==3.1.3
py-cord==2.4.1
PyAutoGUI==0.9.53
pybind11==2.8.0
pycparser==2.21
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
PyNaCl==1.4.0
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pytweening==1.0.7
pytz==2022.4
Quart==0.19.1
Quart-Discord==2.1.4
requests==2.31.0
six==1.16.0
soupsieve==2.5
speedtest-cli==2.1.3
SQLAlchemy==1.4.39
toml==0.10.2
tomli==2.0.1
typing==3.7.4.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
uvicorn==0.23.2
visitor==0.1.3
wavelink==1.3.3
wcwidth==0.2.8
websockets==11.0.3
Werkzeug==3.0.0
wikipedia==1.4.0
wrapt==1.11.0
wsproto==1.2.0
WTForms==3.0.1
yarl==1.9.2
youtube-dl==2021.12.17
zipp==3.8.0
zope.interface==5.4.0
